+++ example/async_/settings.py
@@ -1,0 +1,5 @@
+"""Django settings for example project."""
+
+from ..settings import *  # noqa: F403
+
+ROOT_URLCONF = "example.async_.urls"