+++ docs/changes.rst
@@ -23,6 +23,7 @@ Serializable (don't include in main)
 * Make ``Panel.panel_id`` a class member.
 * Update all panels to utilize data from ``Panel.get_stats()`` to load content
   to render. Specifically for ``Panel.title`` and ``Panel.nav_title``.
+* Extend example app to contain an async version.
 
 Pending
 -------